requirements.txt
flask==2.0.1
flask-cors==3.0.10
flask-login==0.5.0
supabase==1.0.3
python-dotenv==0.19.0
nltk==3.6.3
numpy>=1.21.2
scipy>=1.15.0
scikit-learn>=1.0.2
transformers==4.11.3
bcrypt==3.2.0
requests==2.26.0
pandas>=1.3.3
python-jose[cryptography]==3.3.0
postgrest==0.10.6 